Deprecate <vrs/RecordableId.h> in favor of <vrs/StreamId.h>
Summary:
The VRS APIs were recently updated to use StreamId instead of RecordableId. The behaviors & semantics are the same, it's just a name change.
This series of diffs updates parts of your module to reflect this change. Each of the successive diffs makes a particular change, and each step should only be checked for correctness only. In particular, each step is likely to not update everything, and that's by design, to keep each step simpler/smaller and easier to validate. The goal however, is that the final step brings the module in a very good place.
If you see more further desirable changes, please follow-up as you see fit.

Reviewed By: kiminoue7

Differential Revision: D30949782

fbshipit-source-id: 1391c6f8f57c90aa6387447c1f16b739a94f7ddf