Fix CachedGaiaFileHandler boost dependency issue
Summary:
Unfortunately, CachedGaiaFileHandler uses boost, and even exposes it in the header.
That's a major pain-point, because it means that we expose boost dependencies on a bunch of things, including a bunch of builds modes/platforms that don't support boost.

This diffs limits CachedGaiaFileHandler to the platforms on which we support Gaia (mac/win/linux).

If that doesn't work, I'll remove any reference to boost from the header (force the use of strings), and ifdef the boost code for the platforms that actually support boost in the cpp.

Reviewed By: kiminoue7

Differential Revision: D31152177

fbshipit-source-id: 5cf68b29a1eecb68889803f9371e4f5a39cb42ae