Properly isolate fb and oss implementations of os/Utils.h
Summary: This is the final step to isolate the implementation of some of the os/Utils.h methods. The fb implementation uses system_utils/os, but the oss version is implemented directly in vrs/os, using boost for now, as a reference implementation with no xros or fb server specializations.

Reviewed By: kiminoue7

Differential Revision: D31552485

fbshipit-source-id: e069b4f81c6fc7620763d6ae5c9425f10a7b220f