Fix h265 decode
Summary:
pyvrs calls readRecord against same index multiple times in a row.
When the video codec is h265, this becomes a problem since we need to keep the internal state.
Instead of clearing the lastRecord every time, keep it when the previous read is happened against the exact same record to avoid screwing up the decoder's internal state.

Reviewed By: georges-berenger

Differential Revision: D31594929

fbshipit-source-id: faaa70ec7d5601f9427c6d6272e0f70f0ba6d149