Remove "is_writable" and "default_size_mb" volume attribute from all XROS package manifest files.
Summary:
We are going to remove the "is_writable" and "default_size_mb" volume attribute entirely from the XORS package manifest schema. This is a refactoring change and it does not change any volume configuration of the applications being affected. These volume attributes are not actively used inside XROS.

Context:

Currently, the package manifest schema requires 2 attributes to be specified for a volume. The default size and the max size. It was meant to be used for future dynamic volumes to start from the default size and grow to the max size. Before we had dynamic volumes, the static volumes are are pre-allocated with space equal to the default size and they never grow.

In the previous half, the storage team already implemented the dynamic volumes. The way it works now is that a dynamic volume will create with no actual space committed (default size = 0 always) and it will grow to the max size. There is no need to have 2 separate volume attributes anymore. Hence, we are removing the default_size_mb attribute. Moving forward, all static volumes will pre-allocate the storage space equal to the volume max size specified in the manifest. All dynamic volumes will start with 0 space but eventually can grow to the volume max size specified in the manifest.

As part of this refactoring, we will break this into three steps:

1) Remove all use of default_size_mb in the manifest. Let the volume service to use max size instead of default size for static volume creation.
2) Remove RPC interface related to default volume size in the launcher and the package service.
3) Remove minimum volume size in the block/metadisk layer.

Reviewed By: h-friederich

Differential Revision: D32117995

fbshipit-source-id: f6bd22ac83cf654fd43bab2b50172ea1f40e8ab8