Misc markdown documentation improvements (#17)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/17

When you first discover the VRS code, it can be difficult to understand which code is important, which isn't. The new ReadMe.md files are meant to help open source readers to navigate the source tree more effectively.
This diff also makes tweaks to other documentation files, fixing spelling and presentation errors.

Reviewed By: caprice-j

Differential Revision: D33242900

fbshipit-source-id: eb42902297283b446e1c40da8c3727108c75f996