Add helper function in RecordFileReader to tell if a stream is likely to contain images
Summary:
We frequently need to know if a stream is very likely to contain images. The logic has been duplicated in many places, and this creates a golden version right at the RecordFileReader level, so we stop code duplication.
We will also expose the information in pyvrs, which is were the worst approximation of the logic are made.

Reviewed By: caprice-j

Differential Revision: D33272487

fbshipit-source-id: f0d99d3d86e301e4948734909de1bfcac4462292