Improve randomness of creation ID
Summary:
Every VRS file has a creation ID, meant to be unique. It might be used to identify the file when we cache the file's index. Now and then, we're getting rare reports of cache corruptions, though we don't have a smoking gun (the cache is usually deleted...). So I'm reconsidering how the unique id is generated.
We've used a counter of ns since EPOCH, which has the advantage of telling us when the file was created. The implementation of ns counters isn't providing actual ns accuracy, and is very platform dependent. This fix makes sure that at least 30 bits are really random in that creation id, while preserving the ability to tell when the file was created, albeit with "only" a 1 second precision now, which is more than enough for our purposes.

Reviewed By: caprice-j

Differential Revision: D33340793

fbshipit-source-id: 288b89afe705798da5a07111563584a8343a9119