Backout D34331585 & Make RecordInfo and RecordFileReader available from CurrentReader
Summary:
This diff backs-out D34331585 (https://github.com/facebookresearch/vrs/commit/d24ebc134fb495db0329d672f03028c720dd1a5d), see that diff for details of why.
With MultiRecordFileReader, it becomes more important to get specific context about the record being read during callbacks. By adding a pointer to the RecordInfo of the record being read and a pointer to the reader, we cheaply provide access to more context information to the use cases that need it.

Note: the back-out breaks additional code that already depends on it, and since our CI makes per-diff compares between the state before & after the diff, we're forced to merge the back-out and the replacement changes in one diff. That's bad practice, but with this before-after compare strategy, there is no practical choice.

Differential Revision: D34513145

fbshipit-source-id: 3d05bba8e9964deaea40bf87376b501295064a3d