Add FilteredFileReader
Summary:
FilteredFileReader is a helper that enables reading select parts of a VRS file. It is mostly useful for copy operations that may filter out some records based on timestamp or type, or streams. It can be seen as a specialized iterator that can selectively read a subset of records of a file, so that they can be "processed", whatever that might mean. A simple case might be to copy all the records, as they are, but in a new cleaned-up container, or only copy the first n-seconds of a file, or only some streams, etc.
The copy functionality itself will come in further diffs.

Reviewed By: kiminoue7

Differential Revision: D34290532

fbshipit-source-id: a9f771cdce6e3145e84a2da85f5940b94f775848