Rename iteration methods
Summary:
FilteredFileReader iterate() needs preparation work to work as expected. Unfortunately, not every call sites prepares iterations properly. This diff creates an all-in-one iterateSafe() method, which does what many expected, and which we can use in most places, fixing undiscovered init issues, which could occur when using relative time parameters and streams filters at the same time. This diff also replaces iterate() with iterateAdvanced(), so as to better signal that the method is not meant for most use cases, and requires more care/attention.
By doing the two operations in the same diff, we can ensure that every call site is using the proper API, with fixing if required.

Reviewed By: kiminoue7

Differential Revision: D34745044

fbshipit-source-id: d9eb188dc670c55917c0f6c0c2b9d2d865aeda82