Rename cmake vrs module vrslib
Summary: We're moving to open source a variant of vrstool using the name "vrs" for that tool, so that the most common tool used to manipulate vrs files can simply be named "vrs". So we should not use the "vrs" name for the library in cmake... No functional changes beyond the renaming.

Reviewed By: kiminoue7

Differential Revision: D34839030

fbshipit-source-id: 7b0b1127a459a219b8196ea5c1e4f2ab83816158