Separate time constraints resolution and validation check
Summary:
FilteredFileReader::resolveTimeConstraints() did two things:
- it possibly resolved relative time constraints (in vrstool, when you use +4 for instance, to mean "first data record timestamp +4 seconds") if they were not resolved yet, using the full index of the file
- checked that the time range was valid

This double-functionality has possibly been hidding improper resolution of time ranges. By removing the resolution of time ranges, and making the replacement method only check the validity of the time range, which will verify that we don't have an unresolved relative time range, we will make sure that we don't run into this issue. As we previously updated code that use the class for iteration, we already ensured that we don't need to make the time resolution in all known places, which we double checked in all code-update places.

Reviewed By: kiminoue7

Differential Revision: D34745105

fbshipit-source-id: 846965d2374bedceb10dabcad47b1072ee241a48