Add new vrs cli tool to open source VRS (#26)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/26

This new executable named "vrs" is a command line tool that is a swiss army knife to work with VRS files in a generic manner. this tool doesn't make any assumption about the type of VRS file used, except maybe that the VRS files use RecordFormat conventions to structure records.
This vrs tool is derived from a Meta-only tool named "vrstool" that can't be open sourced as is, because it uses libraries to work with Meta infra, code that can not be open sourced (and wouldn't make any sense outside of Meta anyways). In this first pass, some functionality is still disabled, as we need to open source more code.

Reviewed By: kiminoue7

Differential Revision: D34671069

fbshipit-source-id: 2e98c74ea93ac17ae3cec35fcb4f227f4cee7f6f