add "immediately-available" check to vrs record reader to support interactive use cases
Summary:
for network-backed VRS files, reading frames can sometimes take 0.5+ seconds (and this is blocking due to the need to avoid parallel access of readers).

To support interactive use cases like seeking around a VRS file, it would be nice if we can query a reader to see if a record would be fast to load. For a disk-backed file this is always true. For a manifold-backed file, this is true if the record has already been prefetched and is in the cache.

Reviewed By: georges-berenger

Differential Revision: D35158166

fbshipit-source-id: e705d131414b3620a971610f7f321a2c560b81a7