Better behavior for list and print cli options
Summary:
FilteredFileReader::iterateSafe() computes the range of records to consider when filtering is enabled, then preroll configuration and state records that might have been stripped, but that are essential for RecordFormat & DataLayout record reading. This functionality is very useful when copying files, so you don't lose configuration & state records essential to interpret files, but it also means that list and print operations print records that are not supposed to be printed.
For the list command: simply skip preroll altogether, since printing record's index metadata doesn't require any RecordFormat or DataLayout definitions.
For the print commands: make sure we don't enable printing until after we have prerolled the configuration and state records.

Reviewed By: kiminoue7

Differential Revision: D35584572

fbshipit-source-id: 70436712129ac5d00128dfd997978a719369ab0e