New BufferRecordReader helper class
Summary: To decode data in a background thread, it can be necessary to have a RecordReader object that can read data from a memory buffer, so that record data (image data in particular) can be read in a particular thread that needs to be shared with all the StreamPlayer objects (file objects are not reentrant), then immediately passed to a StreamReader specific background thread for processing. To decode data in that background thread, we now need a RecordReader that can read data from a memory buffer rather than a FileHandler. This is that helper class.

Differential Revision: D36891313

fbshipit-source-id: ffb635a4488a5ee2da20b5278949d64dca585ae8