New JobQueue::waitForJob() method & enabled improvements
Summary:
This new JobQueue::waitForJob() without a timeout method helps simplify a lot of code, as we can now often remove a while loop or an explicit test.
This diff updates all the call sites that could be improved in our code base, which should clearly demonstrate the better code patterns.

Differential Revision: D36917097

fbshipit-source-id: d73ef1296924766b8617af0ee8bc6e5fb3e86d10