New JobQueue::waitForJob() method & enabled improvements (#52)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/52

This new JobQueue::waitForJob() without a timeout method helps simplify a lot of code, as we can now often remove a while loop or an explicit test.
This diff updates all the call sites that could be improved in our code base, which should clearly demonstrate the better code patterns.

Reviewed By: kiminoue7

Differential Revision: D36917097

fbshipit-source-id: 643aa0d3fe2e4f706f2a0e0a02923ae553fad6ab