Update to Qt6 compatible APIs (#56)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/56

To build with Qt6 is OSS, we need to update a few API calls deprecated in Qt6. In most cases, the new code works both with Qt5 and Qt6, but not always. With version checks, the code builds without warning both with Qt5 and Qt6.
We need to preserve many old behaviors, because some of the new methods were added in Qt 5.14, while GitHub CI only get Qt 5.12.

Reviewed By: kiminoue7

Differential Revision: D36997864

fbshipit-source-id: e577c29ad7a49b927343c2c0dff2182ac366de09