Add option to save frames to disk from the UI
Summary:
Make it possible to save images shown in the player by right-clicking on an image and selecting the new option "Save Image As...".
PNG and JPG images are saved as they were in the original VRS file (the original buffer is read from the VRS file again). Video images are saved as displayed, converted to png. They may not have the full resolution of the original data. Raw images a re-read from disk and converted to preserve possible 16 bit depth as png16, but they are converted from whatever format they were in to png.

Note that the context menu should only show the new option when an actual image is visible.

Reviewed By: kiminoue7

Differential Revision: D37095647

fbshipit-source-id: 565352c1a8b7ebf4e4682b900733bbbfce8cd4d9