Custom WriteFileHandler for the body of a VRS file
Summary:
This new unit test demonstrates how to create a custom WriteFileHandler that will handle the user records part of a VRS file, with minimal implementation requirements (no seek, no chunking).
Note that the VRS file generated will have two chunks: the head of the VRS file, created using a regular DiskFile handler at the location specified by the URI, which will contains the description and the index records, but no data records. The second chunk will contain exclusively the user records (configuration, state and data records), which contain the sensor data. With this implementation, it is created next to the VRS head file, with the "_1" prefix.
It is therefore easy to make a custom implementation of the this WriteFileHandler to maybe:
- implement file level encryption (though the file's header will not be encrypted)
- implement an optimized version that can bypass file system APIs and achieve higher performance, maybe writing directly to a device driver than needs no file system buffering.

In order to read the produced VRS file, one will have to have the VRS head file and the "body" file next to each other, or concatenated in a single file, provided the file creation was not interupted.

Reviewed By: kiminoue7

Differential Revision: D37904098

fbshipit-source-id: 58b150a7880250ade951653bfc7b3cf9b9156370