Restore preset named "Default" when opening a file
Summary:
By design vrsplayer reloads the last stream configuration saved with the stream profile of the file loaded.
With this diff, if there is a preset named "Default", it will be automatically loaded instead.
This will make it possible for folks who want the stream configuration to automatically return to a known state when a VRS file is loaded to get there.

Reviewed By: kiminoue7

Differential Revision: D38418204

fbshipit-source-id: c4856df25993c0b2d3e015021d7b9d5ec411880b