Add jpeg-xl support to VRS (empty implementation)
Summary:
This diff add support for jpeg-xl to VRS, but without actual jpeg-xl implementation.
The jpeg-xl library dependencies don't work everywhere, so we can't easily add support for jpeg-xl for everything. This diff adds all the definitions & placeholder locations in the VRS and pyvrs code, but without actual data decoder, to be added later.

Differential Revision: D38682433

fbshipit-source-id: fcbd5a8e317f902fefc459cab1527c711270a5cd