Add jpeg-xl support to VRS (empty implementation) (#67)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/67

This diff add support for jpeg-xl to VRS, but without actual jpeg-xl implementation.
The jpeg-xl library dependencies don't work everywhere, so we can't easily add support for jpeg-xl for everything. This diff adds all the definitions & placeholder locations in the VRS and pyvrs code, but without actual data decoder, to be added later.

Reviewed By: kiminoue7

Differential Revision: D38682433

fbshipit-source-id: f57740b355224bd1a29d07018eab2e22c74bd9e3