Fix python package publish logic to test.pypi.org
Summary:
According to the error message and pypi website, it turns out that test.pypi.org doesn't accept the version form <major>.<minor>.<patch>+<hash>.
Reviewing other internal repository, they are also stripping the hash part.
In this diff, we remove the environmental var used to add the hash to the version number.

Differential Revision: https://internalfb.com/D41758935

fbshipit-source-id: d2b5dd909cf5612542c821b44416b12687e3f2f6