Fix python package publish logic to test.pypi.org (#100)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/100

According to the error message and pypi website, it turns out that test.pypi.org doesn't accept the version form <major>.<minor>.<patch>+<hash>.
Reviewing other internal repository, they are also stripping the hash part.
In this diff, we remove the environmental var used to add the hash to the version number.

Reviewed By: georges-berenger

Differential Revision: D41758935

fbshipit-source-id: 84f6dd3aa95644834f0acd6dd52479c9aa8856ab