Change deploy.yml to only deploy macos for now
Summary:
It turns out VRS doesn't build with manylinux right now with a lot of older version libraries installed by default. It requires us to set up the proper installation scripts for all the library we depend on that take some time.
For now, disable linux build and make the workflow work for macos as a first step.

Reviewed By: georges-berenger

Differential Revision: D41989351

fbshipit-source-id: 1a3ebec2d62f8e6d145aa75e792d551d5fb670ca