Extend reader interface to get a stream by the flavor and index number
Summary:
Adds the functions
`get_stream_for_flavor` to the python API of the vrs library. This function was already present in the C++ version.

I'd like to continue expanding Flavor usage and found this function to be missing. There's another function - `get_streams(RecordableId, Flavor) -> List[StreamIds]` that you could also call. However it gives a list of all of the streamIds (index 0 -> N). With this API exposed, we can now get a single stream without having to parse the list.

Reviewed By: georges-berenger

Differential Revision: D42425169

fbshipit-source-id: 5daca5b12557b6f9bd5567149261144722a87cf6