Add option to only do chunk detection after validating the file's data
Summary: In some situations, we need to only do VRS' link and chunking resolution if we know for sure the file is a VRS file, so as to not mistakenly do that for other file types, for instance, when doing a verbatim upload. This diff adds this option and applies it where needed.

Reviewed By: kiminoue7

Differential Revision: D44590981

fbshipit-source-id: f48c613239c25e052d4a117bfb67a93ba9c8d4a3