Add VRS file stream signature
Summary:
It can be very useful to capture a file "signature", that describes the streams configuration, so we can verify later than a file relates to an original one, with the same stream configuration.
That signature includes stream types, serial numbers and record count per type, but deliberatly does not use stream instance IDs, which might be modified during copies/recompression and other processing that preserve stream structures.

Differential Revision: D45036584

fbshipit-source-id: 8b684e82e08ebd06460c2e837ac3a6772f4fb17c