Rename vrstool option: zero -> zero-vrs and vector -> vector-vrs
Summary:
The variant names are "zero-vrs" and "vector-vrs", so it makes sense to rename the VRStool command line options "--zero" and "--vector" to use those less ambiguous names.
For backward compatibility, in vrstool, we will continue to treat "--zero" like "--zero-vrs". We don't need the same for "vector", because this option not announced yet.

Differential Revision: D45587212

fbshipit-source-id: 8f5beb786c976183deab7176d280d38896b5642f