Add FileDelegator support to FileHandlerFactory
Summary:
FileHandlerFactory is where the logic of FileHandler management lives, and where customization can happen. Now that we've decoupled FileHandler and FileDelegator, we're adding explicit support for FileDelegators to FileHandlerFactory.
Note: because FileDelegator don't have the ability to clone themselves and we can't keep locks while processing delegation,  getFileDelegator() returns a pointer.

Reviewed By: kiminoue7

Differential Revision: D45633507

fbshipit-source-id: af93774f23d61558b91fd9cc43d8c498a99194ec