Add FileSpec extra handling delegation system to FileHandlerFactory, for FileVariant
Summary:
FileVariant needs to be able to process FileSpec based on particular FileSpec extras key/values.
With this diff, FileVariant can now register with FileHandlerFactory a FileDelegator that points to a FileVariant, so extras such as "variant" = "zero-vrs" can be intercepted in FileHandlerFactory, without FileHandlerFactory knowing about the existance of FileVariant.
See how the constant "variant" appears only in FileVariant.

Reviewed By: kiminoue7

Differential Revision: D45644169

fbshipit-source-id: 4de227c46e6b7c10ddb2f0f604d571c2fd8e73a9