Disable automatic creation of variants for large files in the cloud from corp by default
Summary:
We got the feedback that creating variants in the cloud from corp by default can be a problem.
With this diff, we stop creating variants in the cloud from corp by default if the source file is larger than 500 MiB, by providing an optional method that can tell if variant creation should be allowed or not, which can be used to reject the creation of variants from anywhere, or force the creation of variants from anywhere, while having "smart" defaults.
Note that local files variant creation is enabled by default, but just like the other configuration, the defaul behavior can be overriden.
We use a max file size, so that vector file creation from zero files remains likely, even from corp.
This diff cleans-up logging a bit.
For local files, we now create the variant using a temporary file location, to avoid finding a broken variant in the way, should the variant creation be interrupted.

Differential Revision: https://internalfb.com/D46203650

fbshipit-source-id: f9369d320f51126cdd2f95ab110dd2576dd37534