Less comfusing logging for parsing error
Summary:
I was fixing a bug where the `device_to_anchors.csv` was actually missing. After I fixed it, I still read this error log (while the file now exists at the path):
```
[InePoseGraphOpt][INFO]: Loading T3 device to anchor from: "/tmp/123/device_to_anchors.csv"
[FileHandler][ERROR]: Path doesn't exist in uri: /tmp/123/device_to_anchors.csv
[ManifoldUtilsImpl][ERROR]: Not URI, assuming local file: /tmp/123/device_to_anchors.csv
```

Turns out I tunnel visionsed and the log line is only saying the "path string" is not in the "uri string"

Making this log line clearer for the next guy.

Reviewed By: georges-berenger

Differential Revision: D46846967

fbshipit-source-id: 667008bd9de79c008cb495e8b2987be9b15ab869