PixelFrame read png 16 bits as 16 bits.
Summary:
This diff allows to read png with 16 bit per pixel and return them as such in a pixel frame. This is useful for some pipeline storing depth map in mm as uint16_t in a png.

Writing a 16bit BW png is already implemented.

The only non trivial thing here is the call to png_set_swap to go from big endian to little endian.

Reviewed By: georges-berenger

Differential Revision: D47325259

fbshipit-source-id: 03ddc369b8dfd2e85a29935f381870358b118d56