Switch to using rapidjson directly so we can run OSS CI on Ubuntu-latest (#128)
Summary:
We've been using Cereal's internal copy of rapidjson. Unfortunately, for some reason, on GitHub's ubuntu-latest, it doesn't appear to include rapidjson anymore. So it's time to get rid of this odd legacy behavior and use rapidjson directly instead.
The odd thing, is the rapidjson's latest official version, v1.1.0 dates from 2016, and lacks a number of fixes we depend on.
So instead of pulling packages, we need to pull directly from github to a known working hash.

Pull Request resolved: https://github.com/facebookresearch/vrs/pull/128

Reviewed By: SeaOtocinclus

Differential Revision: D47816623

fbshipit-source-id: ae4a6c2adce5f35590afceee7ef1bda320dc3e03