Be more tolerant with missing stride data for the last line of image data
Summary:
Files have been recorded with no stride for the last line, breaking lots of tooling, players in particular. This is too extreme, but we don't want to hide the problem either.
This diff special cases the incomplete or missing stride for the last line of an image, not preventing the image from being read (after all, all the pixel data was read as expected), but logging loudly 500 times, and then only once every 5 seconds. This is a compromise between breaking and hiding problems, being in your face for a reasonable amount of time.

Reviewed By: kiminoue7

Differential Revision: D48410740

fbshipit-source-id: c749a7e1a5a63bca7ee0541d30236f1a749bd020