Add MS-SSIM VRS file compare
Summary:
In order to control compression quality, we need to be able to compare VRS files using an algorithm that can compare image differences and provide a score. The FileImageCompare facility will compare VRS files, using a filtered reader so we don't necessarily have to read every single frame (which can be slow).

meta:
Ocean provides an MS-SSIM implementation, which is an industry standard which we can simply depend on.

Reviewed By: kiminoue7

Differential Revision: D51345482

fbshipit-source-id: 3cc6d9c70b62cc9239ca01e2ea05bb74567620ed