PixelFrame async processing helpers
Summary: When trying to speed up image decoding, we need to do the image processing in a separate than the disk/vrs thread. These new methods allow us to store the image data as read quickly from disk, so that image can be decoded and/or converted in another thread, reducing contention on the thread reading the vrs file.

Reviewed By: kiminoue7

Differential Revision: D51789222

fbshipit-source-id: 44019a7591749b0bebf4eda83b9faf27be02c51e