Use human readable time when logging progress
Summary:
Intead of using the time counter, which we then supplement with a full date/time when we open a file, just make progress logger use a human readable date/time...
While we're at it, let's print the "easy" path, which doesn't reveal gaia paths, which was the motivation for not printing the file's path originaly.
So obvious.

Reviewed By: kiminoue7

Differential Revision: D52435453

fbshipit-source-id: 5304cce1267ad1ac5f24ad887db3d60dd3347cad