Be more conservative when collecting old records
Summary: Because records of different streams may have the same timestamp, when collecting "old" records to write, it's safer to exclude the limit time, so that if a record is later produced with the same timestamp as the limit given, it is more likely to be sorted correctly when written out to disk. This condition can happen when copying files with vrstool.

Reviewed By: kiminoue7

Differential Revision: D52425279

fbshipit-source-id: 166fe9049c5913a238837b5bbebeb598326677b3