SilentLogger: silence duration messages
Summary:
We are getting a lot of noise from the VRS open progress logger, since we open a decent amount of VRS files.

Even if I silence it on the XR Input side, I noticed that I was still getting duration messages. This just makes it so SilentLogger also silences duration messages

Reviewed By: georges-berenger

Differential Revision: D53141868

fbshipit-source-id: f080a464358b8ab7d7176e800dfa862650864167