New AudioBlock helper class to handle audio content blocks
Summary: For images, we have PixelFrame, that helps manage images and image compression. We had nothing equivalent for audio content blocks, which are becoming more complex, with the introduction of Opus compression. By creating this helper class and making vrsplayer use it, we simplify audio handling for vrsplayer and other client apps, that will also want to access uncompressed audio. This class will help extract audio as well.

Differential Revision: D53553128

fbshipit-source-id: a860d8c769b6602668b76fa9a07be9c28c4696ec