Add support for per-record sample count specification via datalayout
Summary:
When decoding audio, in particular compressed audio, we might need to know how many samples are encoded in the audio block to interpret the data.
With this diff, it is now possible to tell in a datalayout preceeding the audio content block how many audio sample frames are coming up.

Reviewed By: kiminoue7

Differential Revision: D53678441

fbshipit-source-id: 7f5d53c5da74b6e646f9f9ed918ab57bcb035fd6