Add use manifold API to source location
Summary:
In order to be able to attribute traffic to Manifold API keys accurately, we need the client code to report which API key it actually used. This code updates the format of the source URI field sent to the backend to include that information using the following format:
- gaia:123456  << when we don't know (yet), eg, when we open a file
- gaia:123456?apikey=the_api_key-key << when we know the API key for a gaia object
- manifold:bucket  << when we don't know (legacy clients)
- manifold:bucket?apikey=the_api_key-key << when we access a file directly in Manifold
- <anything else> << if the data isn't a URI at all, or a URI not starting with "manifold:" or "gaia:", just log that string as is as the source URI. No warnings, no errors.

Note: this diff requires backend updates to handle these, as we're currently getting errors when we add the apikey parameter.

Reviewed By: hanghu

Differential Revision: D56257904

fbshipit-source-id: d031aa8e15b301f0f571be64f162f215128f0b28