Remove crasher caused by sticky lambda
Summary:
By using a lambda that we connect each time we open a file with audio, we create permanent links to objects that aren't permanent, leading to hard to catch non-systematic crashers. By using native Qt connections, we also ensure clean-up.
This issue was introduced in the last vrsplayer diff.

Reviewed By: finik

Differential Revision: D56736500

fbshipit-source-id: 9a464d1723b2bb9fc33ed17f26a88cd5311ab73d