Add stream semantic convention tags
Summary:
Add stream tag convention definitions to specify a semantic for the images in a stream, and provide hints on how to render such images.
This can help us better render depth images as grey8 images, and optimize scaling.
This makes it possible to interpret grey16 images as segmentation images, to be rendered in rgb8 with unique visually distinct colors for each value.

Reviewed By: liorda

Differential Revision: D57598125

fbshipit-source-id: cf1d2ef86d8ccc33118d226810a0b0e1f23b4d85