Add image semantic to streams based on RecordableTypeId [RFC]
Summary:
We have a new stream tag convention that lets us specify how to interpret some image data, but we also have a huge number of existing files that don't use these new tags.
With this diff, we hardcode a few RecordableTypeId values and stream flavors with image semantics, so we can better interpret old files.

Reviewed By: liorda

Differential Revision: D57598705

fbshipit-source-id: ad4a7e5a1c7d757bfd3c3faef866786e1fcebf2e