vrs: pass fileSpec.extras as options to DiskFile.create
Summary:
This allows `WriteFileHandler` objects to make use of URI parameters that they do not save off at `checkChunks` time. (Like `DiskFile`).

Not that some `WriteFileHandler`s just preserve the entire URI in the `chunk.path` field, allowing them to access these later. `DiskFile` notably does NOT, which is why this is necesary.

Reviewed By: georges-berenger

Differential Revision: D57869385

fbshipit-source-id: c4fa7c4d9972545cbd6c7a62d74835b31e488cde