fix isRecordAvailableOrPrefetch
Summary:
There was a bug in `isRecordAvailableOrPrefetch` where we were testing for availability of a slighty too big bytes range.

The size of a vrs record includes the header, as described in
https://www.internalfb.com/code/fbsource/[7baffe629cac]/arvr/libraries/vrs/vrs/FileFormat.h?lines=151

Hence, we should remove the length of the header when testing for the record availability. This is because the reader has been read just before in the same function.

Reviewed By: georges-berenger

Differential Revision: D58656057

fbshipit-source-id: 2ead7bf268f4d06348036cab9d57253ab228c412