Print segmentation colors and with object class / object ID
Summary: Print in a terminal the list of class/object IDs used when rendering segmentation data, with color sample, and in the case of object class, the category. To trigger the print, use the new "Print Segmentation Colors" menu option, with the shortcut Cmd-P on Mac, and Ctrl-P on Windows.

Reviewed By: liorda

Differential Revision: D58003683

fbshipit-source-id: e68b71a17cf6e7f40e8a8bcbe55b0f3de14e8f30