Reduce jxl multithreading
Summary: We observed that aggressive multithreading could lead to increased psnr in compressed images, while also increasing overall CPU costs. This diff removes the psnr bump from observed test cases, and reduce the overall default CPU overhead, though it does increase wall time a bit, which is a trade-off we might need to tune and/or provide some control for with decompression.

Reviewed By: hanghu

Differential Revision: D59707832

fbshipit-source-id: 12d69ed5b7c6148226a6d3c6fe14d5daba0ac1ce