Add option to not fully sort records by timestamp when generating a file
Summary:
If we want to maximize write IO when creating a VRS file with compression, we need to avoid any wait on compresion threads. If there is no compression at all, then we're already never waiting to write records as fast as possible. But if we compress records, we want to avoid the situation when the next record to write isn't compressed yet, while others are ready to write. Note that this only makes sense when writing to a local file with diskfile or asyncdiskfile. By default, vrs records are always sorted.

To disable record sorting, specify the output path with the URI parameter "sort_records=false", like so:
  diskfile:/path/to/file.vrs?sort_records=false
or even:
  /path/to/file.vrs?sort_records=false

Note: if you use this option, please make sure to always process the file with records sorting enabled before uploading the resulting file to the cloud, to avoid streaming inneficiencies because of making the streaming cache much less effective.

Reviewed By: jtbraun

Differential Revision: D60749758

fbshipit-source-id: ac92f70385ec7f5b62cd2390228e16bfacdaa31f