Sort records fully for checksums
Summary:
When reading a VRS file, records are sorted by timestamp, streamID, and position in the file, but not by record type, which means that two files can have the same records, but in slightly different order, and not be equivalent for checksums. As we added the `sort_records` option to generate file that might not be fully sorted, it's easy to generate file that are equivalent, but have different checksums.
When opening a file, you can now specify the same `sort_records` options to get the file fully sorted, taking into account the record's type, to get a deterministic state.

Reviewed By: kiminoue7

Differential Revision: D60942282

fbshipit-source-id: cdf94109f29ea794f0cedafc0655e38873123045