Provide method to create records with a DataSource part that isn't copied
Summary: In order to do high speed realtime data collection, we need to avoid copying large images. This diff adds the option to create uncompressed records, with a part that isn't copied when the record object is created, allowing to write the data directly in the output file when the record is written to do.

Reviewed By: markhog

Differential Revision: D62920807

fbshipit-source-id: ead63ad190e649a55e4da1c7eaa410f9ee2f25fb