Add debayering normalization for PixelFormat::BAYER8_RGGB
Summary:
BAYER8_RGGB is currently displayed as if it was a grey8 image, which is very approximative. This diff adds a meta-only code path in which we debayer these images.
Because debayering very large images is slow, we use a debayering + downsampler algorithm for display, but with large images only.

Reviewed By: maxouellet

Differential Revision: D63151489

fbshipit-source-id: 119a047d4a0e362daa1c033caf8c90abacbc3468