New tool to export information about every image in a VRS file [POC]
Summary:
This is  a POC API + experimentation/debug CLI tool.
The tool demonstrate that we can extract the information necessary to read every image in a VRS file without going through a RecordFileReader. Use the API and tool in the next diff to use the data to extract frames.

To be clear: the tool is just for experimentation, and maybe debugging purposes. For production use, we expect to use only the API, in C++ or through a Python bindings TBD.

Differential Revision: D64584016

fbshipit-source-id: da200d7414e2d9f2be842cc91e157020912b45d3